Move MusixmatchLyricsGetter to appropriate folder
