+++ app/src/main/java/com/github/fribourgsdp/radio/external/musixmatch/MusixmatchLyricsGetter.kt
@@ -1,10 +1,8 @@
 package com.github.fribourgsdp.radio.external.musixmatch
 
-import android.util.Log
 import com.github.fribourgsdp.radio.util.JSONParser
 import com.github.fribourgsdp.radio.util.JSONStandardParser
 import okhttp3.*
-import org.json.JSONArray
 import java.io.IOException
 import java.util.*
 import java.util.concurrent.CompletableFuture
@@ -20,18 +18,21 @@ const val QUERY_ARTIST_FIELD = "q_artist"
 const val SORT_CONDITION = "s_artist_rating=desc"
 
 interface LyricsGetter{
-    fun getLyrics(songName: String,artistName: String = "",client: OkHttpClient = OkHttpClient(),parser: JSONParser = JSONStandardParser()): CompletableFuture<String>
+    fun getLyrics(songName: String, artistName: String = "", client: OkHttpClient = OkHttpClient(), parser: JSONParser = JSONStandardParser()): CompletableFuture<String>
     fun getSongID(songName: String, artistName: String, client: OkHttpClient = OkHttpClient(), parser : JSONParser = JSONStandardParser()) : CompletableFuture<Int>
-    fun markSongName(lyrics : String, name : String) : String
+    fun markSongName(lyrics: String, name: String): String
 }
 /**
  * Tool to get lyrics from a given song name and artist using Musixmatch API.
  * API Call doc : https://stackoverflow.com/questions/45219379/how-to-make-an-api-request-in-kotlin
  */
 object MusixmatchLyricsGetter : LyricsGetter {
-    const val LYRICS_NOT_FOUND = "---No lyrics were found for this song.---"
+    const val LYRICS_NOT_FOUND = "---No lyrics were found for this song. Feel free to add yours !---"
+    const val BACKEND_ERROR_PLACEHOLDER = ""
 
-    class LyricsNotFoundException(val reason : String = LYRICS_NOT_FOUND) : Exception()
+    abstract class LyricsGetterException(val default : String) : Exception()
+    class NoLyricsFoundForThisSong : LyricsGetterException(LYRICS_NOT_FOUND)
+    class BackendError : LyricsGetterException("")
 
     /**
      * Asks Musixmatch and retrieves the lyrics of a song.
@@ -47,21 +48,27 @@ object MusixmatchLyricsGetter : LyricsGetter {
         client: OkHttpClient,
         parser: JSONParser
     ): CompletableFuture<String> {
-        Log.println(Log.ASSERT, "*", "LYRICS GETTER CALL !!!!")
+//        Log.println(Log.ASSERT, "*", "LYRICS GETTER CALL !!!!")
         val future = CompletableFuture<String>()
         val trackIDFuture = getSongID(songName, artistName, client)
-        val trackID: Int
-        try {
-            trackID = trackIDFuture.get()
-        } catch (e: Throwable) {
-            future.completeExceptionally(LyricsNotFoundException())
-            return future
+        trackIDFuture.handle { trackID, e1 ->
+            if (e1 != null) {
+                future.complete((e1 as LyricsGetterException).default)
+            } else {
+                val url = "$BASE_URL$TRACK_LYRICS_GET?$TRACK_ID_FIELD=$trackID&$API_KEY_FIELD=$API_KEY"
+                val request = Request.Builder().url(url).build()
+                val future2 = CompletableFuture<String>()
+                client.newCall(request).enqueue(GetLyricsCallback(future2, parser))
+                future2.whenComplete { s, e2 ->
+                    if (e2 != null) {
+                        future.complete((e2 as LyricsGetterException).default)
+                    } else {
+                        future.complete(cleanLyrics(s))
+                    }
+                }
+            }
         }
-        val url = "$BASE_URL$TRACK_LYRICS_GET?$TRACK_ID_FIELD=$trackID&$API_KEY_FIELD=$API_KEY"
-
-        val request = Request.Builder().url(url).build()
-        client.newCall(request).enqueue(GetLyricsCallback(future, parser))
-        return future.thenApply { s -> cleanLyrics(s) }
+        return future
     }
 
     /**
@@ -84,27 +91,25 @@ object MusixmatchLyricsGetter : LyricsGetter {
 
     private class GetLyricsCallback(private val future : CompletableFuture<String>, private val parser : JSONParser) : Callback {
         override fun onFailure(call: Call, e: IOException) {
-            future.completeExceptionally(e)
+            future.completeExceptionally(BackendError())
         }
 
         override fun onResponse(call: Call, response: Response) {
             val parsedResponseString = response.body()?.string()
             val parsedResponse = parser.parse(parsedResponseString)
-            val lyrics : String?
+            var lyrics : String?
             if(parsedResponse == null){
-                future.completeExceptionally(LyricsNotFoundException())
+                future.completeExceptionally(BackendError())
             } else {
                 val status = parsedResponse.getJSONObject("message").getJSONObject("header").getInt("status_code")
-                if (status == 404) {
-                    future.completeExceptionally(LyricsNotFoundException())
-                } else {
+                checkStatus(status, future) {
                     lyrics = parsedResponse
                         .getJSONObject("message")
                         .getJSONObject("body")
                         .getJSONObject("lyrics")
                         .getString("lyrics_body")
                     if (lyrics?.isEmpty() == true) {
-                        future.completeExceptionally(LyricsNotFoundException())
+                        future.completeExceptionally(NoLyricsFoundForThisSong())
                     } else {
                         future.complete(lyrics)
                     }
@@ -115,30 +120,30 @@ object MusixmatchLyricsGetter : LyricsGetter {
 
     private class GetSongIDCallback(private val future : CompletableFuture<Int>, private val parser : JSONParser) : Callback {
         override fun onFailure(call: Call, e: IOException) {
-            future.completeExceptionally(e)
+            future.completeExceptionally(BackendError())
         }
 
         override fun onResponse(call: Call, response: Response) {
             val parsedResponse = parser.parse(response.body()?.string())
             if(parsedResponse == null){
-                future.completeExceptionally(Exception("Error parsing response"))
-            }
-            val trackList = try {
-                parsedResponse
-                    ?.getJSONObject("message")
-                    ?.getJSONObject("body")
-                    ?.getJSONArray("track_list")
-            } catch (e : Exception){
-                JSONArray()
-            }
-            if(trackList?.length() == 0){
-                future.completeExceptionally(LyricsNotFoundException())
-            } else {
-                val firstTrackID = trackList
-                    ?.getJSONObject(0)
-                    ?.getJSONObject("track")
-                    ?.getInt("track_id")
-                future.complete(firstTrackID)
+                future.completeExceptionally(BackendError())
+            } else{
+                val status = parsedResponse.getJSONObject("message").getJSONObject("header").getInt("status_code")
+                checkStatus(status, future) {
+                    val trackList = parsedResponse
+                        .getJSONObject("message")
+                        .getJSONObject("body")
+                        .getJSONArray("track_list")
+                    if (trackList.length() == 0) {
+                        future.completeExceptionally(NoLyricsFoundForThisSong())
+                    } else {
+                        val firstTrackID = trackList
+                            .getJSONObject(0)
+                            ?.getJSONObject("track")
+                            ?.getInt("track_id")
+                        future.complete(firstTrackID)
+                    }
+                }
             }
         }
     }
@@ -148,7 +153,7 @@ object MusixmatchLyricsGetter : LyricsGetter {
      * The mention to Musixmatch will be displayed elsewhere, like in the activity displaying the lyrics.
      */
     private fun cleanLyrics(lyrics : String) : String{
-        if(lyrics == LYRICS_NOT_FOUND) {
+        if(lyrics == LYRICS_NOT_FOUND || lyrics == "") {
             return lyrics
         }
         val allLines = lyrics.split("\n")
@@ -170,5 +175,25 @@ object MusixmatchLyricsGetter : LyricsGetter {
             lyrics
                 .replace(name, "<strike>${name[0].uppercase() + name.lowercase().substring(1)}</strike>", ignoreCase = true)
                 .replace("\n", "<br>")
+    }
+
+    /**
+     * If status is 404, it means this song will never have lyrics.
+     * If status is not 404 but still of the form 4xx or 5xx, this request could have encountered an error, and another request could work
+     * Else if no problem execute function.
+     */
+    private fun <T> checkStatus(
+        status: Int,
+        future: CompletableFuture<T>,
+        function: () -> Boolean, ) {
+        when {
+            status == 404 -> {
+                future.completeExceptionally(NoLyricsFoundForThisSong())
+            }
+            status >= 400 -> {
+                future.completeExceptionally(BackendError())
+            }
+            else -> function.invoke()
+        }
     }
 }
